+++ changelog.md
@@ -8,6 +8,7 @@
 
 ### Changes
 - Updated to Forge 1.18.1-39.0.64
+- Blocks dropped by vein miner, tree miner, and 3x3 mining now drop their resources on the player
 
 ----------------------- 0.5.2 Changes -----------------------
 ### Additions